Expandable items (#11)
* feat(covid-list): Fix data

Remove unused data class, update map, add logging interceptor

* feat(covid-list): Add expandable drawable

Add drawable

* feat(covid-list): Add expandable list

Updated list entity, optimized with payloads

* feat(covid-list): Refactored list

Removed unused classes

* feat(covid-list): List animations

Added list expand animations

* feat(covid-list): Editor config

Added .editorconfig

* feat(covid-list): Editor config

Added max-line-length. Reformatted code

* feat(covid-list): Reformat code

Reformatted code

+++ covid-data/src/main/java/com/moshenskyi/feature_covid_data/data/model/vaccine/VaccineCapacityInfoItem.kt
@@ -2,5 +2,5 @@ package com.moshenskyi.feature_covid_data.data.model.vaccine
 
 data class VaccineCapacityInfoItem(
     val country: String,
-    val timeline: Map<String, Int>
+    val timeline: Map<String, Long>
 )
